refactor: 덱 관련 패키지 위치 변경
