+++ src/main/kotlin/blackjack/domain/deck/CardPattern.kt
@@ -1,4 +1,4 @@
-package blackjack.domain
+package blackjack.domain.deck
 
 enum class CardPattern {
     SPADE,
